+++ package/telegram_client/example/example.dart
@@ -38,7 +38,7 @@ import 'package:general_lib/general_lib.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:path/path.dart';
 import 'package:telegram_client/scheme/telegram_client_library_tdlib_option_parameter.dart';
- import 'package:telegram_client/telegram_client.dart';
+import 'package:telegram_client/telegram_client.dart';
 import 'package:universal_io/io.dart';
 
 void main(List<String> args) async {