+++ package/telegram_client/examples/tdlib.dart
@@ -16,8 +16,12 @@ void main(List<String> args) async {
     clientOption: TelegramClientLibraryTdlibOptionParameter.create(
       // api_id: 0,
       // api_hash: "",
-      database_directory: Directory(path.join(Directory.current.uri.toFilePath(), "temp", "db")).path,
-      files_directory: Directory(path.join(Directory.current.uri.toFilePath(), "temp", "file")).path,
+      database_directory:
+          Directory(path.join(Directory.current.uri.toFilePath(), "temp", "db"))
+              .path,
+      files_directory: Directory(
+              path.join(Directory.current.uri.toFilePath(), "temp", "file"))
+          .path,
     ),
   );
   await tdlib.ensureInitialized();
@@ -28,7 +32,8 @@ void main(List<String> args) async {
     if (update["@type"] == "updateAuthorizationState") {
       if (update["authorization_state"] is Map) {
         Map authorization_state = update["authorization_state"];
-        if (authorization_state["@type"] == "authorizationStateWaitPhoneNumber") {
+        if (authorization_state["@type"] ==
+            "authorizationStateWaitPhoneNumber") {
           Map res = await tdlib.invoke(
             "setAuthenticationPhoneNumber",
             parameters: {
@@ -67,7 +72,8 @@ void main(List<String> args) async {
   for (var i = 1; i <= 1000; i++) {
     print(i);
     await Future.delayed(Duration(microseconds: 10));
-    Directory directory = Directory(path.join(Directory.current.path, "temp", "lp_${i}"));
+    Directory directory =
+        Directory(path.join(Directory.current.path, "temp", "lp_${i}"));
     Map res = await tdlib.createclient(
       clientId: tdlib.td_create_client_id(),
       clientOption: TelegramClientLibraryTdlibOptionParameter.create(
@@ -77,7 +83,8 @@ void main(List<String> args) async {
     );
     res.printPretty();
   }
-  print(DateTime.now().extension_general_lib_countAgoFromDateTime(dateTime: dateTime));
+  print(DateTime.now()
+      .extension_general_lib_countAgoFromDateTime(dateTime: dateTime));
   stdin.listen((e) async {
     try {
       DateTime dateTime = DateTime.now();
@@ -98,7 +105,8 @@ void main(List<String> args) async {
       } else {
         return;
       }
-      print(DateTime.now().extension_general_lib_countAgoFromDateTime(dateTime: dateTime));
+      print(DateTime.now()
+          .extension_general_lib_countAgoFromDateTime(dateTime: dateTime));
     } catch (e) {
       print(e);
     }