+++ package/telegram_client/lib/tdlib/tdlib_library/base.dart
@@ -114,9 +114,11 @@ abstract class TdlibBaseCore {
 ///
 abstract class TdlibBase implements TdlibBaseCore {
   final ReceivePort receivePort = ReceivePort();
-  final TelegramClientLibraryTdlibOptionParameter client_option = TelegramClientLibraryTdlibOptionParameter.create(
+  final TelegramClientLibraryTdlibOptionParameter client_option =
+      TelegramClientLibraryTdlibOptionParameter.create(
     api_id: num.tryParse("OTQ1NzU=".general_lib_utils_decryptFromBase64()),
-    api_hash: 'YTM0MDZkZThkMTcxYmI0MjJiYjZkZGYzYmJkODAwZTI='.general_lib_utils_decryptFromBase64(),
+    api_hash: 'YTM0MDZkZThkMTcxYmI0MjJiYjZkZGYzYmJkODAwZTI='
+        .general_lib_utils_decryptFromBase64(),
     database_directory: "tg_db",
     files_directory: "tg_file",
     use_file_database: true,
@@ -127,7 +129,8 @@ abstract class TdlibBase implements TdlibBaseCore {
     system_language_code: 'en',
     new_verbosity_level: 0,
     application_version: 'v1',
-    device_model: 'VGVsZWdyYW0gQ2xpZW50IEFaS0FERVYgR0xPQkFMIENPUlBPUkFUSU9O'.general_lib_utils_decryptFromBase64(),
+    device_model: 'VGVsZWdyYW0gQ2xpZW50IEFaS0FERVYgR0xPQkFMIENPUlBPUkFUSU9O'
+        .general_lib_utils_decryptFromBase64(),
     system_version: Platform.operatingSystemVersion,
     database_key: "",
     start: true,
@@ -154,9 +157,12 @@ abstract class TdlibBase implements TdlibBaseCore {
   Duration invoke_time_out = Duration(minutes: 10);
   double timeOutUpdate;
   bool is_invoke_throw_on_error = false;
-  FutureOr<void> Function(dynamic update, TdlibBase libTdJson)? on_receive_update;
-  FutureOr<String> Function(int client_id, TdlibBase libTdJson)? on_generate_extra_invoke;
-  FutureOr<Map> Function(String extra, int client_id, TdlibBase libTdJson)? on_get_invoke_data;
+  FutureOr<void> Function(dynamic update, TdlibBase libTdJson)?
+      on_receive_update;
+  FutureOr<String> Function(int client_id, TdlibBase libTdJson)?
+      on_generate_extra_invoke;
+  FutureOr<Map> Function(String extra, int client_id, TdlibBase libTdJson)?
+      on_get_invoke_data;
   int task_max_count;
   int task_min_cooldown;
   TdlibBase({
@@ -210,12 +216,15 @@ abstract class TdlibBase implements TdlibBaseCore {
         TdlibIsolateReceiveData tdlibIsolateReceiveData = update;
         try {
           if (tdlibIsolateReceiveData.updateData["@extra"] is String) {
-            event_emitter.emit(eventName: event_invoke, value: tdlibIsolateReceiveData);
+            event_emitter.emit(
+                eventName: event_invoke, value: tdlibIsolateReceiveData);
           } else {
-            event_emitter.emit(eventName: event_update, value: tdlibIsolateReceiveData);
+            event_emitter.emit(
+                eventName: event_update, value: tdlibIsolateReceiveData);
           }
         } catch (e) {
-          event_emitter.emit(eventName: event_update, value: tdlibIsolateReceiveData);
+          event_emitter.emit(
+              eventName: event_update, value: tdlibIsolateReceiveData);
         }
       } else if (update is TdlibIsolateReceiveDataError) {
         is_init_isolate = false;
@@ -334,7 +343,9 @@ abstract class TdlibBase implements TdlibBaseCore {
           client_new_option[key] = value;
         }
       });
-      final Map<String, num> tdlib_option_should_not_empty_num = {"api_id": client_option.api_id ?? 0};
+      final Map<String, num> tdlib_option_should_not_empty_num = {
+        "api_id": client_option.api_id ?? 0
+      };
       tdlib_option_should_not_empty_num.forEach((key, value) {
         try {
           if (client_new_option[key] is num == false) {
@@ -507,8 +518,11 @@ abstract class TdlibBase implements TdlibBaseCore {
     required String extra,
     required bool isAutoGetChat,
     required bool isInvokeThrowOnError,
-    required FutureOr<String> Function(int client_id, TdlibBase libTdJson)? onGenerateExtraInvoke,
-    required FutureOr<Map> Function(String extra, int client_id, TdlibBase libTdJson)? onGetInvokeData,
+    required FutureOr<String> Function(int client_id, TdlibBase libTdJson)?
+        onGenerateExtraInvoke,
+    required FutureOr<Map> Function(
+            String extra, int client_id, TdlibBase libTdJson)?
+        onGetInvokeData,
   }) async {
     final Completer<Map> completer = Completer<Map>();
     late final EventEmitterListener listener;
@@ -530,7 +544,8 @@ abstract class TdlibBase implements TdlibBaseCore {
             if (extra.isEmpty) {
               parameters["@extra"] = extra;
             } else if (onGenerateExtraInvoke != null) {
-              parameters["@extra"] = await onGenerateExtraInvoke(clientId, this);
+              parameters["@extra"] =
+                  await onGenerateExtraInvoke(clientId, this);
             } else {
               parameters["@extra"] = generateUuid(15);
             }
@@ -543,7 +558,9 @@ abstract class TdlibBase implements TdlibBaseCore {
           return extra_procces;
         }();
 
-        if (isAutoGetChat && RegExp(r"^(sendMessage|getChatMember)$", caseSensitive: false).hashData(parameters["@type"])) {
+        if (isAutoGetChat &&
+            RegExp(r"^(sendMessage|getChatMember)$", caseSensitive: false)
+                .hashData(parameters["@type"])) {
           if (parameters["chat_id"] is int) {
             td_send(clientId, {
               "@type": "getChat",
@@ -627,7 +644,9 @@ abstract class TdlibBase implements TdlibBaseCore {
       event_emitter.off(listener: listener);
     } catch (e) {}
 
-    if (result.isEmpty || result["@type"] is String == false || result["@type"] == "error") {
+    if (result.isEmpty ||
+        result["@type"] is String == false ||
+        result["@type"] == "error") {
       if (isInvokeThrowOnError) {
         result["@type"] = "error";
         throw result;
@@ -660,8 +679,10 @@ abstract class TdlibBase implements TdlibBaseCore {
     String? extra,
     bool? isAutoGetChat,
     bool? isInvokeThrowOnError,
-    FutureOr<String> Function(int client_id, TdlibBase libTdJson)? onGenerateExtraInvoke,
-    FutureOr<Map> Function(String extra, int client_id, TdlibBase libTdJson)? onGetInvokeData,
+    FutureOr<String> Function(int client_id, TdlibBase libTdJson)?
+        onGenerateExtraInvoke,
+    FutureOr<Map> Function(String extra, int client_id, TdlibBase libTdJson)?
+        onGetInvokeData,
   }) async {
     isUseCache ??= false;
     durationCacheExpire ??= Duration(
@@ -703,8 +724,10 @@ abstract class TdlibBase implements TdlibBaseCore {
     bool? isUseCache,
     Duration? durationCacheExpire,
     bool? isAutoGetChat,
-    FutureOr<String> Function(int client_id, TdlibBase libTdJson)? onGenerateExtraInvoke,
-    FutureOr<Map> Function(String extra, int client_id, TdlibBase libTdJson)? onGetInvokeData,
+    FutureOr<String> Function(int client_id, TdlibBase libTdJson)?
+        onGenerateExtraInvoke,
+    FutureOr<Map> Function(String extra, int client_id, TdlibBase libTdJson)?
+        onGetInvokeData,
     bool? isInvokeThrowOnError,
   }) async {
     return await invoke(