+++ package/telegram_client_isar_scheme/CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.8.23
+## 0.8.24
 
-- Update: 2024-08-30 14:33:46.231053 - update librarry.
+- Update: 2024-09-24 02:31:41.965876 - update library.
+- Revisi And Change Style Code for improve MAX speed